pyproject.toml
[project]
name = "cover-letter-agent-v2"
version = "0.1.0"
description = "AI agent using Google ADK that analyzes resumes and creates personalized cover letters with comprehensive research"
authors = [
    { name = "Cover Letter Agent V2", email = "agent@example.com" },
]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.0.0"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
], version = "^1.93.0" }
google-genai = "^1.9.0"
google-generativeai = "^0.8.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
pypdf2 = "^3.0.1"
pymupdf = "^1.24.0"
requests = "^2.32.0"
beautifulsoup4 = "^4.12.0"
openai = "^1.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"